Code for creating api structure
CW provides a JSON file describing the API structure, which can be used to set up picklists, explore the API, etc. However, we don't want to parse it all out every time we use the module, so we save the parsed info to an xml file that can be quickly imported.